pyproject.toml
[project]
name = "decision-prism"
version = "0.1.0"
description = "Dynamic Strategic Decision System — AI-powered debate and analysis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.12.0",
    "rich>=13.0",
    "tavily-python>=0.3.0",
    "firecrawl-py>=1.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "jinja2>=3.1.0",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
    "nest-asyncio>=1.6.0",
]

[project.scripts]
decision-prism = "decision_prism.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "respx>=0.21.0",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["decision_prism*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
